requirements.txt
langchain_google_genai 
langchain-community
assemblyai[extras]
pyaudio
fastapi
uvicorn[standard]
aiohttp
python-dotenv
requests